+++ assembly/serialize/index/arbitrary.ts
@@ -7,6 +7,7 @@ import { serializeInteger } from "./integer";
 import { serializeMap } from "./map";
 import { serializeObject } from "./object";
 import { serializeString } from "./string";
+import { serializeDynamic } from "./typedarray";
 
 export function serializeArbitrary(src: JSON.Value): void {
   switch (src.type) {
@@ -59,6 +60,10 @@ export function serializeArbitrary(src: JSON.Value): void {
       break;
     case JSON.Types.Map:
       serializeMap(src.get<Map<string, JSON.Value>>());
+      break;
+    case JSON.Types.TypedArray:
+    case JSON.Types.ArrayBuffer:
+      serializeDynamic(src.type, src.get<usize>());
       break;
     default: {
       const fn = JSON.Value.METHODS.get(src.type - JSON.Types.Struct);